pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personalizer"
version = "0.1.0"
description = "Minimalist terminal personal dashboard — next-hour focus, learning, and reminders."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Arpit Yadav" }]
dependencies = [
    "textual>=0.80",
    "httpx>=0.27",
    "google-api-python-client>=2.120",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "openai>=1.0",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "freezegun>=1.5",
    "pytest-textual-snapshot>=0.4",
    "textual-dev>=1.5",
]

[project.scripts]
personalizer = "personalizer.__main__:main"
personalizer-setup = "personalizer.scripts.gcal_setup:main"

[tool.hatch.build.targets.wheel]
packages = ["src/personalizer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
